Adding code to watch ShardReplication in discovery, just refactoring existing code with a callback function.
